requirements.txt
anthropic==0.8.1
SQLAlchemy==2.0.27
PyYAML==6.0.1
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
pytest==8.0.0
pytest-asyncio==0.23.2
black==23.12.1
isort==5.13.2
mypy==1.8.0 